feat(ui): give the bottle action bar one primary control
Open C: Drive, Terminal, Winetricks and Run were four identical flat labels on
a material strip, so opening a terminal looked as important as launching a
game. Run is the reason the app exists and is now the only prominent control
on the bar.

The bar moves to its own view while it is being touched, since BottleView was
already at the type length limit.

Liquid Glass where the system has it: GlassEffectContainer is what lets the
capsules blend rather than stack their own layers, and it is macOS 26 only, so
the older path is the same buttons in bordered styles. The emphasis lands
either way, which is the part that matters.
